requirements.txt
Keras
python-socketio
requests
tensorflow
